Move controls.rs to psynth::control module and break into smaller files pot.rs, mux.rs, flow.rs for better scoping -- TODO: decide if leaving everything nested like this (psynth::control::pot::sine_pot) is preferable over a flatter organization, (e.g. psynth::sine_pot or psynth::control::sine_pot)
