+++ src/control/mod.rs
@@ -1,0 +1,3 @@
+pub mod pot;
+pub mod mux;
+pub mod flow;
