--- src/controls.rs
@@ -1,283 +1,0 @@
-use std::cell::{RefCell, Cell};
-use std::sync::{mpsc, Arc, Mutex};
-use std::thread;
-
-use anyhow::Result;
-
-use crate::{generators, filters, Pot, Generator, FilterComposable, Sample};
-
-
-/// Allow for the usage of raw floats as `f32` potentiometers when control over the value is not
-/// necessary.
-impl Pot<f32> for f32 {
-    fn read(&self) -> f32 {
-        *self
-    }
-}
-
-
-impl Pot<f64> for f64 {
-    fn read(&self) -> f64 {
-        *self
-    }
-}
-
-
-/// Modulate the value returned via the held `Generator`.
-///
-/// Useful for composing `Generators` as `Pot` inputs to other `Generator`s or `Filter`s.
-pub struct GeneratorPot {
-    gen: RefCell<Generator>,
-}
-
-impl GeneratorPot {
-    pub fn new(generator: Generator) -> Self {
-        Self {
-            gen: RefCell::new(generator),
-        }
-    }
-}
-
-impl Pot<f32> for GeneratorPot {
-    fn read(&self) -> f32 {
-        (&mut *self.gen.borrow_mut())()
-    }
-}
-
-
-/// Sinusoidally osciallating potentiometer.
-///
-/// Crank back and forth at the provided `frequency`, oscillating between the provided `low` and
-/// `high` values.
-pub fn sine_pot<P>(sample_rate: u32, frequency: P, low: f32, high: f32) -> GeneratorPot
-where
-    P: Pot<f32> + 'static
-{
-    GeneratorPot::new(
-        generators::sine(sample_rate, frequency)
-            .compose(filters::offset(1.0))
-            .compose(filters::gain((high - low) / 2.0))
-            .compose(filters::offset(low)))
-}
-
-
-/// Interactively read values from stdin via `readline`.
-pub struct StdinPot<T> {
-    cur: Cell<T>,
-    receiver: mpsc::Receiver<T>,
-}
-
-
-impl Default for StdinPot<f32> {
-    fn default() -> Self {
-        Self::new("f32pot", 0.0, |line| Ok(line.parse::<f32>()?))
-    }
-}
-
-
-impl Default for StdinPot<f64> {
-    fn default() -> Self {
-        Self::new("f64pot", 0.0, |line| Ok(line.parse::<f64>()?))
-    }
-}
-
-
-impl<T> StdinPot<T>
-where
-    T: Send + 'static,
-{
-    /// Create a new `StdinPot`, spawning a stdin reader thread.
-    pub fn new<F>(name: &str, default: T, converter: F) -> Self
-    where
-        F: Fn(&str) -> Result<T> + Send + 'static,
-    {
-        let prompt = format!("{}> ", name);
-        let mut reader = rustyline::Editor::<()>::new();
-        let (sender, receiver) = mpsc::channel();
-        thread::spawn(move || loop {
-            match reader.readline(prompt.as_str()) {
-                Ok(l) if l == "q" => {
-                    println!("exit requested, exiting gracefully...");
-                    std::process::exit(0);
-                },
-                Ok(l) => {
-                    match converter(l.as_str()) {
-                        Ok(val) => {
-                            sender.send(val).unwrap();  // thread panic if channel is closed
-                            reader.add_history_entry(l.as_str());
-                        },
-                        Err(e) => eprintln!("unable to parse '{}', try again (reason: {:?})", l, e),
-                    }
-                },
-                Err(e) => {
-                    eprintln!("readline error: {:?}", e);
-                    std::process::exit(1);
-                },
-            }
-        });
-        Self {
-            cur: Cell::new(default),
-            receiver: receiver,
-        }
-    }
-}
-
-
-impl<T> Pot<T> for StdinPot<T>
-where
-    T: Send + Copy + 'static,
-{
-    fn read(&self) -> T {
-        match self.receiver.try_recv() {
-            Ok(val) => {
-                self.cur.set(val);
-                val
-            },
-            Err(_) => self.cur.get(),
-        }
-    }
-}
-
-
-/// Enables sharing of a `Pot` impl in multiple places.
-impl<T, P> Pot<T> for Arc<Mutex<P>>
-where
-    T: Send + Copy + 'static,
-    P: Pot<T> + 'static,
-{
-    fn read(&self) -> T {
-        let inner = self.lock().unwrap();
-        (*inner).read()
-    }
-}
-
-
-impl<T, F> Pot<T> for F
-where
-    T: Send + Copy + 'static,
-    F: Fn() -> T + Send + 'static,
-{
-    fn read(&self) -> T {
-        (self)()
-    }
-}
-
-
-/// Mux together two left/right streams in a stereo setup using a custom mux function.
-///
-/// The muxing is performed by the provided `mux_function` that determines how much of each
-/// channels' signal should contribute to a channel's output at any given sample.
-///
-/// The yielded `Generator`s are entangled in that calling one also calls the other. This is
-/// important to take note of for `Generator` implementations that keep some sort of internal state.
-pub fn mux2<F>(
-    mux_function: F,
-    left: Generator,
-    right: Generator,
-) -> (Generator, Generator)
-where
-    F: FnMut(Sample, Sample) -> (Sample, Sample) + Send + 'static,
-{
-    // TODO: use single Arc for whole shared state, instead of 3 Arcs for the three components
-    // that are shared?
-    let vals_left = Arc::new(Mutex::new((None::<Sample>, None::<Sample>)));
-    let vals_right = Arc::clone(&vals_left);
-
-    let generators_left = Arc::new(Mutex::new((left, right)));
-    let generators_right = Arc::clone(&generators_left);
-
-    let mux_f_left = Arc::new(Mutex::new(mux_function));
-    let mux_f_right = Arc::clone(&mux_f_left);
-
-    let out_left: Generator = Box::new(move || {
-        let mut vals_unlocked = vals_left.lock().unwrap();
-        match *vals_unlocked {
-            (Some(_), Some(_)) => unreachable!("neither value collected -- should never occur"),
-            (Some(l), None) => {
-                *vals_unlocked = (None, None);
-                l
-            },
-            (None, _) => {
-                let (ref mut left_gen, ref mut right_gen) = &mut *generators_left.lock().unwrap();
-                let ref mut mux_f = &mut *mux_f_left.lock().unwrap();
-                let (l, r) = mux_f(left_gen(), right_gen());
-                *vals_unlocked = (None, Some(r));
-                l
-            },
-        }
-    });
-
-    // this would get tedious for more than two channels -- is there a general-form solution for
-    // this multi-stream muxing problem?
-    let out_right: Generator = Box::new(move || {
-        let mut vals_unlocked = vals_right.lock().unwrap();
-        match *vals_unlocked {
-            (Some(_), Some(_)) => unreachable!("neither value collected -- should never occur"),
-            (None, Some(r)) => {
-                *vals_unlocked = (None, None);
-                r
-            },
-            (_, None) => {
-                let (ref mut left_gen, ref mut right_gen) = &mut *generators_right.lock().unwrap();
-                let ref mut mux_f = &mut *mux_f_right.lock().unwrap();
-                let (l, r) = mux_f(left_gen(), right_gen());
-                *vals_unlocked = (Some(l), None);
-                r
-            },
-        }
-    });
-
-    (out_left, out_right)
-}
-
-
-/// Fork the provided `Generator` into two entangled `Generator`s that will yield the same value
-/// on each at any given instant.
-pub fn fork(generator: Generator) -> (Generator, Generator) {
-    mux2(move |l, _| (l, l), generator, generators::silence())
-}
-
-
-/// Join the provided `Generator` streams into a single `Generator`.
-///
-/// Allows composition of multiple input sources. Serves a similar purpose for `Generator`s as
-/// `filters::parallel` serves for `Filter`s.
-pub fn join(mut generators: Vec<Generator>) -> Generator {
-    Box::new(move || {
-        let mut out = 0f32;
-        for generator in generators.iter_mut() {
-            out += generator();
-        }
-        out
-    })
-}
-
-
-/// Join the two `Generator`s into a single `Generator`.
-///
-/// The `bias` potentiometer determines how much of each signal contributes to the end result.
-/// A value of `≤-1` is 100% `left`, `≥1` is 100% `right`, and 0 is an even 50%/50% split.
-pub fn join2<P>(bias: P, mut left: Generator, mut right: Generator) -> Generator
-where
-    P: Pot<f32> + 'static,
-{
-    let mut clipper = filters::clip(-1.0, 1.0);
-    Box::new(move || {
-        let l_val = left();
-        let r_val = right();
-        let b_val = clipper(bias.read()) + 1.0;
-        ((2.0 - b_val) * 0.5 * l_val) + (b_val * 0.5 * r_val)
-    })
-}
-
-
-pub fn balance<P>(
-    _balance_pot: P,
-    _left: Generator,
-    _right: Generator,
-) -> (Generator, Generator)
-where
-    P: FnMut(Sample, Sample) -> (Sample, Sample) + Send + 'static,
-{
-    unimplemented!()
-}
