+++ src/lib.rs
@@ -3,7 +3,7 @@ pub mod generators;
 pub mod filters;
 pub mod consumers;
 pub mod observers;
-pub mod controls;
+pub mod control;
 pub mod sampling;
 pub mod keys;
 pub mod devices;
@@ -92,7 +92,7 @@ impl FilterComposable for Generator {
     where
         F: FnMut(Generator, Generator) -> Generator + Send,
     {
-        let (left_gen, right_gen) = controls::fork(self);
+        let (left_gen, right_gen) = control::flow::fork(self);
         join_function(left_gen, right_gen)
     }
 }
